+++ applications/_dendritic_solidification/custom_pde.h
